refactor: 번호 생성 전략 패키지 분리 및 리팩토링
